Testing prime faces (#13)
* Merge dev into Juan

Conflicts:
	.gitignore
	build/web/WEB-INF/classes/Hwp07/Credentials.sql
	build/web/WEB-INF/classes/Hwp07/Customer.class
	build/web/WEB-INF/classes/Hwp07/RegisteredBean.class
	build/web/customerFolder/profile.xhtml
	build/web/index.xhtml
	build/web/regForm.xhtml
	nbproject/private/private.properties
	nbproject/private/private.xml
	src/java/Hwp07/Credentials.sql
	src/java/Hwp07/Customer.java
	src/java/Hwp07/RegisteredBean.java
	web/index.xhtml

* Added Primefaces, it created welcomePrimefaces.xhtml.

* In collectionview.xhtml:
	Changed to <p:datatable>.
	Changed to <p:column>.
	Changed to <p:link>.
	Changed to <p:commandButton>.
	Added resizeableColumns.
	Added pagination with 10 rows per page.
	Added dragableColumns.

* In collectionview.xhtml:
	Added paginatorTemplate.
	Added rowsPerPageTemplate with 10, 25, 50, 100 rows per page.
	Added sortBuy for each column.

* In editcollection.xhtml:
	Changed to <p:datatable>.
	Changed to <p:column>.
	Changed to <p:link>.
	Changed to <p:commandButton>.
	Added resizeableColumns.
	Added pagination with 10 rows per page.
	Added dragableColumns.
	Added paginatorTemplate.
	Added rowsPerPageTemplate with 10, 25, 50, 100 rows per page.
	Added sortBuy for each column.
Changed "In Collection" to "Copies" in editcollection.xhtml and collectionview.xhtml.

* Started DataTable Edit functionality in editcollection.xhtml which requires editing both editcollection.xhtml and the EditCollectionBean. Unfortunately this is not currently working. I'm just commiting my work.

* In editcollection.xhtml when an album is deleted the form is now refreshed with ajax. In both editcollection.xhtml and collectionview.xhtml the refresh button refreshes the form with ajax.

* Created AddAlbumBean to separate add album functionality from EditCollectionBean to accommodate future changes to editcollection.xhtml.
Added more albums to Album.sql.

* Added Album List header to editcollection.xhtml and collectionview.xhtml.

* Merge dev into Juan

Conflicts:
	.gitignore
	build/web/WEB-INF/classes/Hwp07/Credentials.sql
	build/web/WEB-INF/classes/Hwp07/Customer.class
	build/web/WEB-INF/classes/Hwp07/RegisteredBean.class
	build/web/customerFolder/profile.xhtml
	build/web/index.xhtml
	build/web/regForm.xhtml
	nbproject/private/private.properties
	nbproject/private/private.xml
	src/java/Hwp07/Credentials.sql
	src/java/Hwp07/Customer.java
	src/java/Hwp07/RegisteredBean.java
	web/index.xhtml

* Merge dev into Juan

Conflicts:
	.gitignore
	build/web/WEB-INF/classes/Hwp07/Credentials.sql
	build/web/WEB-INF/classes/Hwp07/Customer.class
	build/web/WEB-INF/classes/Hwp07/RegisteredBean.class
	build/web/customerFolder/profile.xhtml
	build/web/index.xhtml
	build/web/regForm.xhtml
	nbproject/private/private.properties
	nbproject/private/private.xml
	src/java/Hwp07/Credentials.sql
	src/java/Hwp07/Customer.java
	src/java/Hwp07/RegisteredBean.java
	web/index.xhtml

* Remove Save Edit button from collectionview.xhtml.

* Modified updateAlbum

+++ src/java/JavaObjects/Album.java
@@ -12,7 +12,7 @@ public class Album {
     private int releaseYear;
     private int numberOfTracks;
     private int numberOfDiscs;
-    private long albumID;
+    private int albumID;
     private int albumCount;
     private boolean editable;
 
@@ -64,11 +64,11 @@ public class Album {
         this.numberOfDiscs = numberOfDiscs;
     }
 
-    public long getAlbumID() {
+    public int getAlbumID() {
         return albumID;
     }
 
-    public void setAlbumID(long albumID) {
+    public void setAlbumID(int albumID) {
         this.albumID = albumID;
     }
